rutas reservas populate

+++ routes/reserva.routes.js
@@ -1,49 +1,36 @@
-const express = require("express");
+const express = require('express');
+const Reserva = require('../models/reserva.model'); 
 const router = express.Router();
+const { verifyToken } = require('../middleware/auth.middleware'); 
 
-const Reserva = require("../models/reserva.model");
-
-const { verifyToken} = require("../middleware/auth.middleware.js");
-
-// GET todas las reservas
-router.get("/", async (req, res, next) => {
+// GET "/api/reservas" - Obtener las reservas del usuario autenticado
+router.get("/", verifyToken, async (req, res, next) => {
   try {
-    const response = await Reserva.find();
-    res.status(200).json(response);  // Cambiado a 200 (OK)
+    const userId = req.payload._id; 
+
+    const reservas = await Reserva.find({ userId }).populate('alojamiento', 'image description hotelName address');
+
+    res.status(200).json(reservas); 
   } catch (error) {
     next(error);
   }
 });
 
-// GET detalles de una reserva específica por ID
-router.get("/:reservaId", async (req, res, next) => {
-  try {
-    const response = await Reserva.findById(req.params.reservaId);
-    if (!response) {
-      return res.status(404).json({ message: "Reserva no encontrada" });
-    }
-    res.status(200).json(response);
-  } catch (error) {
-    next(error);
-  }
-});
 
 // POST nueva reserva
+
 router.post("/addReserva", verifyToken, async (req, res, next) => {
   try {
+    console.log(req.payload);
     const response = await Reserva.create({
       guestName: req.body.guestName,
-      userId: req.payload._id,  // Tomar el ID del usuario autenticado
-      image: req.body.image,
-      description: req.body.description,
-      phone: req.body.phone,
-      hotelName: req.body.hotelName,
-      address: req.body.address,
-      roomNumber: req.body.roomNumber,
+      userId: req.payload._id,  
+      alojamiento: req.body.alojamiento,  // el ID del alojamiento
       checkInDate: req.body.checkInDate,
       checkOutDate: req.body.checkOutDate,
       numberOfGuests: req.body.numberOfGuests
     });
+
     res.status(201).json(response);
   } catch (error) {
     next(error);
@@ -53,10 +40,20 @@ router.post("/addReserva", verifyToken, async (req, res, next) => {
 // DELETE eliminar una reserva específica
 router.delete("/:reservaId", verifyToken, async (req, res, next) => {
   try {
-    const response = await Reserva.findByIdAndDelete(req.params.reservaId);
-    if (!response) {
+  
+    const reserva = await Reserva.findById(req.params.reservaId);
+
+    if (!reserva) {
       return res.status(404).json({ message: "Reserva no encontrada" });
     }
+
+    if (reserva.userId.toString() !== req.payload._id) {
+      return res.status(403).json({ message: "No tienes permiso para eliminar esta reserva" });
+    }
+
+    // Eliminar la reserva
+    await Reserva.findByIdAndDelete(req.params.reservaId);
+    
     res.status(200).json({ message: "Reserva eliminada correctamente" });
   } catch (error) {
     next(error);
@@ -70,12 +67,7 @@ router.put("/:reservaId/edit", verifyToken, async (req, res, next) => {
       req.params.reservaId,
       {
         guestName: req.body.guestName,
-        image: req.body.image,
-        description: req.body.description,
-        phone: req.body.phone,
-        hotelName: req.body.hotelName,
-        address: req.body.address,
-        roomNumber: req.body.roomNumber,
+        userId: req.payload._id,  
         checkInDate: req.body.checkInDate,
         checkOutDate: req.body.checkOutDate,
         numberOfGuests: req.body.numberOfGuests
